pyproject.toml
[tool.poetry]
name = "figus"
version = "0.1.0"
description = ""
authors = ["Eric Kersten <ekersten@gmail.comn>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.1.1"
djangorestframework = "^3.13.1"

[tool.poetry.dev-dependencies]
flake8 = "^5.0.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
